requirements.txt
gradio
pandas
numpy
requests
beautifulsoup4
python-dotenv
konlpy
# LangGraph and related libraries
langgraph
langchain_core
# Selenium for advanced scraping
selenium
webdriver-manager
# LangChain for LLM calls
langchain-google-genai
# For plotting
matplotlib
# For Word document generation
python-docx
